Update classgroup.tex
The sentence that I edited was previously correct since the intersection $\p\cap \Z$ is going to be prime even when $\p$ is the zero ideal, but the equality $\p\cap \Z = p\Z$ in this case would force $p=0$, which doesn't coincide with the definition of prime numbers. I therefore changed the sentence slightly in an attempt to be clearer about that.